add placesControllers directory with atomic controllers for async requests
