+++ src/controllers/placesControllers/getPlace.js
@@ -1,0 +1,28 @@
+const getPlace = async (req, res) => {
+  const { id } = req.params
+
+  try {
+    const singlePlace = await place.getSinglePlace(id)
+
+    res.json({
+      message: 'success',
+      payload: {
+        data: singlePlace,
+        description: 'Place found',
+        statusCode: 200,
+      },
+    })
+  } catch (err) {
+    console.error(err)
+    res.json({
+      message: 'failure',
+      error: {
+        err,
+        description: 'Place not found.',
+        statusCode: 404,
+      },
+    })
+  }
+}
+
+export { getPlace }
